tried making the delay work but something goes wrong in the write function, i get a segmentation fault

+++ CSD2c/1.les/1.opdracht/circularBuffer.h
@@ -7,20 +7,20 @@ class CircularBuffer{
         CircularBuffer(float size, float numSamplesDelay);
         ~CircularBuffer();
 
-        void setReadIndex(int numSamplesDelay);    
+        void setReadIndex(float numSamplesDelay);
         float read();
-        void write(float sample);  
+        void write(float sample);
 
     private:
         int wrap(int head);
-        int size; 
-        //size of the buffer, depending on how long i want the delay to be 
+        int size;
+        //size of the buffer, depending on how long i want the delay to be
         int numSamplesDelay;
-        //numSamplesDelay, space between readIndex and writeIndex 
-        int readIndex;
-        //read the samples out of the buffer 
+        //numSamplesDelay, space between readIndex and writeIndex
+        unsigned int readIndex;
+        //read the samples out of the buffer
         int writeIndex;
-        //write the info of the samples in the buffer 
+        //write the info of the samples in the buffer
         float* buffer;
 
         //the readIndex an writeIndex are always equally far apart = numSamplesDelay
